Split example into different files
