+++ gb-dbg/examples/debug_eframe/memory.rs
@@ -1,0 +1,11 @@
+pub struct Memory {
+    pub memory: Vec<u8>,
+}
+
+impl Default for Memory {
+    fn default() -> Self {
+        Self {
+            memory: vec![0xFFu8; u16::MAX as usize],
+        }
+    }
+}
